Get rid of app actions, move everything into the catalog domain
