+++ package.json
@@ -171,6 +171,7 @@
     "@types/react-transition-group": "^4.4.0",
     "@types/react-virtualized": "^9.21.10",
     "@types/react-window": "^1.8.2",
+    "@types/redux-mock-store": "^1.0.2",
     "@types/semver-compare": "^1.0.1",
     "@types/serialize-javascript": "^4.0.0",
     "@types/superagent": "^4.1.10",
