+++ src/stores/appcatalog/constants.ts
@@ -22,3 +22,9 @@ export const PREPARE_INGRESS_TAB_DATA_ERROR = 'PREPARE_INGRESS_TAB_DATA_ERROR';
 export const CATALOG_LOAD_INDEX_REQUEST = 'CATALOG_LOAD_INDEX_REQUEST';
 export const CATALOG_LOAD_INDEX_SUCCESS = 'CATALOG_LOAD_INDEX_SUCCESS';
 export const CATALOG_LOAD_INDEX_ERROR = 'CATALOG_LOAD_INDEX_ERROR';
+
+export const CLUSTER_LOAD_APP_README_REQUEST =
+  'CLUSTER_LOAD_APP_README_REQUEST';
+export const CLUSTER_LOAD_APP_README_SUCCESS =
+  'CLUSTER_LOAD_APP_README_SUCCESS';
+export const CLUSTER_LOAD_APP_README_ERROR = 'CLUSTER_LOAD_APP_README_ERROR';
