+++ src/stores/global/types.ts
@@ -1,22 +1,29 @@
 import {
+  CLUSTER_SELECT,
   GLOBAL_LOAD_ERROR,
   GLOBAL_LOAD_REQUEST,
   GLOBAL_LOAD_SUCCESS,
 } from 'stores/global/constants';
 
-export interface GlobalLoadRequestAction {
+export interface IGlobalLoadRequestAction {
   type: typeof GLOBAL_LOAD_REQUEST;
 }
 
-export interface GlobalLoadSuccessAction {
+export interface IGlobalLoadSuccessAction {
   type: typeof GLOBAL_LOAD_SUCCESS;
 }
 
-export interface GlobalLoadErrorAction {
+export interface IGlobalLoadErrorAction {
   type: typeof GLOBAL_LOAD_ERROR;
 }
 
+export interface IGlobalSelectClusterAction {
+  type: typeof CLUSTER_SELECT;
+  clusterID: string;
+}
+
 export type GlobalActions =
-  | GlobalLoadRequestAction
-  | GlobalLoadSuccessAction
-  | GlobalLoadErrorAction;
+  | IGlobalLoadRequestAction
+  | IGlobalLoadSuccessAction
+  | IGlobalLoadErrorAction
+  | IGlobalSelectClusterAction;
